--- language/en-GB/en-GB.plg_jshoppingadmin_wt_on_fly_image_handler.ini
@@ -1,15 +1,0 @@
-PLG_WT_ON_FLY_IMAGE_HANDLER="JShoppingadmin - WT On fly image handler"
-PLG_WT_ON_FLY_IMAGE_HANDLER_DESC="Processing of JoomShopping product images during editing: the same proportions, conversion to the same format, removal of transparency, naming of files. The Intervention Image library is used."
-PLG_WT_ON_FLY_IMAGE_HANDLER_SQUARE_IMAGE="Make images square?"
-PLG_WT_ON_FLY_IMAGE_HANDLER_SQUARE_IMAGE_YES="The images will be reduced to square proportions based on the largest side of the image. The extra empty background will be cut off, the necessary fields will be added to the square proportion, the background will be filled with white."
-PLG_WT_ON_FLY_IMAGE_HANDLER_SQUARE_IMAGE_NO="The image sizes will be changed according to the JoomShopping settings. Excess areas of the image are cut off. If the values of both height and width are specified, the larger side will be changed to the one specified in the settings, the smaller side will be changed proportionally. If only one value is specified - height or width - the second side will be changed proportionally. If the height and width in the settings are 0, the image will simply be saved in the specified format without resizing."
-PLG_WT_ON_FLY_IMAGE_HANDLER_OUTPUT_FILE_FORMAT="Output file format"
-PLG_WT_ON_FLY_IMAGE_HANDLER_ONBEFOREEDITCONFIGIMAGE_NOTICE="<div class=\"alert alert-info\"><h3>Attention!</h3><p>Image processing when saving the product is carried out by the plugin <a class=\"alert-link\" href=\"index.php?option=com_plugins&filter.search=WT%20On%20fly%20image%20handler\" target=\"_blank\">JShoppingadmin - WT On fly image handler</a>. When processing, the settings of the image size and the quality of the processed images are taken into account.</p><p>The settings \"Resizing\" and \"Image fill color\" are ignored.</p></div>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_RENAME_IMAGES="Rename image files?"
-PLG_WT_ON_FLY_IMAGE_HANDLER_RENAME_IMAGES_DESC="Image file name will be human-friendly (product name + digital index)."
-
-PLG_WT_ON_FLY_IMAGE_HANDLER_DOWNLOAD_JINTERVENTION = "<a href=\"https://hika.su/update/free/lib_jinterventionimage.zip\" class=\"btn btn-success\">Download JIntervention Image library [1]</a> <a href=\"https://web-tolk.ru/dev/biblioteki/jintervention-image.html\" class=\"btn btn-success\">Download JIntervention Image library [2]</a>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_ENABLE_JINTERVENTION = "<a href=\"index.php?option=com_installer&view=manage&filter.search=jintervention\" target=\"_blank\" class=\"btn btn-warning\">Enable JIntervention Image library</a>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_ENABLED="JIntervention Image enabled"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_DISABLED="JIntervention Image <span class=\"label label-important\">disabled</span>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_NOT_FOUND="JIntervention Image <span class=\"label label-important\">not found</span>"
