--- language/en-GB/en-GB.plg_jshoppingadmin_wt_on_fly_image_handler.sys.ini
@@ -1,15 +1,0 @@
-PLG_WT_ON_FLY_IMAGE_HANDLER="JShoppingadmin - WT On fly image handler"
-PLG_WT_ON_FLY_IMAGE_HANDLER_DESC="Processing of JoomShopping product images during editing: the same proportions, conversion to the same format, removal of transparency, naming of files. The Intervention Image library is used."
-PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_INSTALL = "Thank you for intalling extension"
-PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_DISCOVER_INSTALL = "Thank you for intalling extension"
-PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_UPDATE = "Thank you for updating extension"
-PLG_WT_ON_FLY_IMAGE_HANDLER_MAYBE_INTERESTING="<h4>It might be interesting</h4><ul class=\"nav nav-tabs nav-stacked list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomshopping/action-log-joomshopping.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Action log - JoomShopping - free plugin.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomshopping/wt-sm-ozon-rocket-dlya-joomshopping.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">WT SM Ozon Rocket for JoomShopping, shipping method.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/joomshopping-russian-post-shipping-method-api.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">JoomShopping Russian Post Shipping method (via API)</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-plugins/wt-joomshopping-bitrix24-pro.html\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">WT JoomShopping Bitrix24 PRO, two-way integration plugin for JoomShopping online store and Bitrix24 CRM.</a></li></ul>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_WHATS_NEW="<h3>v.1.0.2 What's new?</h3><p>Joomla 4 and JoomShopping 5 support.</p><h4>Demo video (Russian language)</h4><iframe width=\"560\" height=\"315\" src=\"https://www.youtube.com/watch?v=6_XOuLCy2ng\" title=\"YouTube video player\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture\" allowfullscreen></iframe></p>"
-
-PLG_WT_ON_FLY_IMAGE_HANDLER_DOWNLOAD_JINTERVENTION = "<a href=\"https://hika.su/update/free/lib_jinterventionimage.zip\" class=\"btn btn-success\">Download JIntervention Image library [1]</a> <a href=\"https://web-tolk.ru/dev/biblioteki/jintervention-image.html\" class=\"btn btn-success\">Download JIntervention Image library [2]</a>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_ENABLE_JINTERVENTION = "<a href=\"index.php?option=com_installer&view=manage&filter.search=jintervention\" target=\"_blank\" class=\"btn btn-warning\">Enable JIntervention Image library</a>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_ENABLED="JIntervention Image enabled"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_DISABLED="JIntervention Image <span class=\"badge badge-important bg-danger\">disabled</span>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_NOT_FOUND="JIntervention Image <span class=\"badge badge-important bg-danger\">not found</span>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_FAILED="The Intervention Image library could not be downloaded and installed automatically. Please install it manually."
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_OK="JIntervention Image library has been installed automatically."
