+++ language/en-GB/plg_jshoppingadmin_wt_on_fly_image_handler.sys.ini
@@ -1,0 +1,19 @@
+PLG_WT_ON_FLY_IMAGE_HANDLER = "JShoppingadmin - WT On fly image handler"
+PLG_WT_ON_FLY_IMAGE_HANDLER_DESC = "Processing of JoomShopping product images during editing: the same proportions, conversion to the same format, removal of transparency, naming of files. The Intervention Image library is used."
+PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_INSTALL = "Thank you for intalling extension"
+PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_DISCOVER_INSTALL = "Thank you for intalling extension"
+PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_UPDATE = "Thank you for updating extension"
+PLG_WT_ON_FLY_IMAGE_HANDLER_MAYBE_INTERESTING = "<h4>It may be interesting</h4><ul class=\"nav nav-tabs nav-stacked list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/web-assets\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla 4 & Joomla 5 Web Assets.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomshopping\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">JoomShopping extensions</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-plugins\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla plugins</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-modules\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla modules</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/components\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla components</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/en/dev/joomla-libraries\" target=\"_blank\" title=\"Go to https://web-tolk.ru\">Joomla libraries for developers.</a></li></ul>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_WHATS_NEW = "<h3>v.2.0.0 What's new?</h3><p>Joomla 4 and JoomShopping 5 support.</p>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JOOMLARU_TELEGRAM_CHAT = "Joomla community in Telegram"
+PLG_WT_ON_FLY_IMAGE_HANDLER_WEBTOLK_TELEGRAM_CHANNEL = "@WebTolkRu in Telegram"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ERROR_COMPATIBLE_PHP = "This version is compatible only with PHP %s and later"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ERROR_COMPATIBLE_JOOMLA = "This version is compatible only with Joomla %s and later"
+
+PLG_WT_ON_FLY_IMAGE_HANDLER_DOWNLOAD_JINTERVENTION = "<a href=\"https://hika.su/update/free/lib_jinterventionimage.zip\" class=\"btn btn-success\">Download JIntervention Image library [1]</a> <a href=\"https://web-tolk.ru/dev/biblioteki/jintervention-image\" class=\"btn btn-success\">Download JIntervention Image library [2]</a>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ENABLE_JINTERVENTION = "<a href=\"index.php?option=com_installer&view=manage&filter.search=jintervention\" target=\"_blank\" class=\"btn btn-warning\">Enable JIntervention Image library</a>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_ENABLED = "JIntervention Image enabled"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_DISABLED = "JIntervention Image <span class=\"badge bg-danger\">disabled</span>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_NOT_FOUND = "JIntervention Image <span class=\"badge bg-danger\">not found</span>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_FAILED = "The Intervention Image library could not be downloaded and installed automatically. Please install it manually."
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_OK = "JIntervention Image library has been installed automatically."
