+++ language/ru-RU/plg_jshoppingadmin_wt_on_fly_image_handler.ini
@@ -1,0 +1,14 @@
+PLG_WT_ON_FLY_IMAGE_HANDLER = "JShoppingadmin - WT On fly image handler"
+PLG_WT_ON_FLY_IMAGE_HANDLER_DESC = "Обработка изображений товара JoomShopping при редактировании: одинаковые пропорции, конвертация в один формат, удаление прозрачности, именование файлов. Используется библиотека Intervention Image."
+PLG_WT_ON_FLY_IMAGE_HANDLER_SQUARE_IMAGE = "Делать изображения квадратными?"
+PLG_WT_ON_FLY_IMAGE_HANDLER_SQUARE_IMAGE_YES = "Изображения будут приведены к квадратным пропорциям исходя из наибольшей стороны изображения. Будет обрезан лишний пустой фон, добавятся необходимые поля до квадратной пропорции, заливка фона белым цветом."
+PLG_WT_ON_FLY_IMAGE_HANDLER_SQUARE_IMAGE_NO = "Размеры изображений будут изменены согласно настройкам JoomShopping. Лишние области изображения отрезаются. Если указаны значения и высоты и ширины - большая сторона будет изменена до указанной в настройках, меньшая сторона будет изменена пропорционально. Если указана только одно значение - высоты или ширины - вторая сторона будет изменена пропорционально. Если высота и ширина в настройках равны 0 - изображение просто будет сохранено в указанном формате без изменения размера."
+PLG_WT_ON_FLY_IMAGE_HANDLER_OUTPUT_FILE_FORMAT = "Выходной формат файла"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ONBEFOREEDITCONFIGIMAGE_NOTICE = "<div class=\"alert alert-info\"><h3>Обратите внимание!</h3><p>Обработка изображений при сохранении товара ведётся плагином <a class=\"alert-link\" href=\"index.php?option=com_plugins&filter.search=WT%20On%20fly%20image%20handler\" target=\"_blank\">JShoppingadmin - WT On fly image handler</a>. При обработке учитываются настройки размера изображений и качества обработанных изображений.</p><p>Настройки \"Изменение размера\" и \"Цвет заливки изображения\" игнорируются.</p></div>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_RENAME_IMAGES = "Переименовывать файлы изображений?"
+PLG_WT_ON_FLY_IMAGE_HANDLER_RENAME_IMAGES_DESC = "Имя файла будет не абракадаброй, а названием товара + цифровой индекс"
+PLG_WT_ON_FLY_IMAGE_HANDLER_DOWNLOAD_JINTERVENTION = "<a href=\"https://hika.su/update/free/lib_jinterventionimage.zip\" class=\"btn btn-success\">Скачать библиотеку JIntervention Image [1]</a> <a href=\"https://web-tolk.ru/dev/biblioteki/jintervention-image.html\" class=\"btn btn-success\">Скачать библиотеку JIntervention Image [2]</a>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ENABLE_JINTERVENTION = "<a href=\"index.php?option=com_installer&view=manage&filter.search=jintervention\" target=\"_blank\" class=\"btn btn-warning\">Включить библиотеку JIntervention Image</a>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_ENABLED = "JIntervention Image включена"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_DISABLED = "JIntervention Image <span class=\"label label-important\">выключена</span>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_NOT_FOUND = "JIntervention Image <span class=\"label label-important\">не установлена</span>"
