+++ language/ru-RU/plg_jshoppingadmin_wt_on_fly_image_handler.sys.ini
@@ -1,0 +1,19 @@
+PLG_WT_ON_FLY_IMAGE_HANDLER = "JShoppingadmin - WT On fly image handler"
+PLG_WT_ON_FLY_IMAGE_HANDLER_DESC = "Обработка изображений товара JoomShopping при редактировании: одинаковые пропорции, конвертация в один формат, удаление прозрачности, именование файлов. Используется библиотека Intervention Image."
+PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_INSTALL = "Спасибо за установку расширения"
+PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_DISCOVER_INSTALL = "Спасибо за установку расширения"
+PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_UPDATE = "Спасибо за обновление расширения"
+PLG_WT_ON_FLY_IMAGE_HANDLER_MAYBE_INTERESTING = "<h4>Может быть интересно</h4><ul class=\"nav nav-tabs nav-stacked list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/web-assets\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Joomla 4 & Joomla 5 Web Assets</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Расширения для JoomShopping</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Плагины Joomla</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-modules\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Модули Joomla</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/components\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Компоненты Joomla</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/biblioteki\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Библиотеки Joomla для разработчиков</a></li></ul>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_WHATS_NEW = "<h3>v.2.0.0 Что нового?</h3><h4>Поддержка Joomla 5</h4><p>Плагин переписан на новую архитектуру расширений Joomla 4 / Joomla 5. Это означает, что он будет работать и в Joomla 6.</p><h4>Удалена поддержка Joomla 3</h4><p>Окончательно удалена поддержка Joomla 3.</p>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JOOMLARU_TELEGRAM_CHAT = "Joomla сообщество в Телеграм"
+PLG_WT_ON_FLY_IMAGE_HANDLER_WEBTOLK_TELEGRAM_CHANNEL = "@WebTolkRu в Telegram"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ERROR_COMPATIBLE_JOOMLA = "Эта версия совместима только с Joomla %s и выше."
+PLG_WT_ON_FLY_IMAGE_HANDLER_ERROR_COMPATIBLE_PHP = "Эта версия совместима только с PHP %s и выше."
+
+PLG_WT_ON_FLY_IMAGE_HANDLER_DOWNLOAD_JINTERVENTION = "<a href=\"https://hika.su/update/free/lib_jinterventionimage.zip\" class=\"btn btn-success\">Скачать библиотеку JIntervention Image [1]</a> <a href=\"https://web-tolk.ru/dev/biblioteki/jintervention-image.html\" class=\"btn btn-success\">Скачать библиотеку JIntervention Image [2]</a>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_ENABLE_JINTERVENTION = "<a href=\"index.php?option=com_installer&view=manage&filter.search=jintervention\" target=\"_blank\" class=\"btn btn-warning\">Включить библиотеку JIntervention Image</a>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_ENABLED = "JIntervention Image включена"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_DISABLED = "JIntervention Image <span class=\"badge bg-danger\">выключена</span>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_NOT_FOUND = "JIntervention Image <span class=\"badge bg-danger\">не установлена</span>"
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_FAILED = "Не удалось скачать и установить библиотеку JIntervention Image автоматически. Пожалуйста, установите её вручную."
+PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_OK = "Библиотека JIntervention Image установлена автоматически."
