--- language/ru-RU/ru-RU.plg_jshoppingadmin_wt_on_fly_image_handler.sys.ini
@@ -1,15 +1,0 @@
-PLG_WT_ON_FLY_IMAGE_HANDLER="JShoppingadmin - WT On fly image handler"
-PLG_WT_ON_FLY_IMAGE_HANDLER_DESC="Обработка изображений товара JoomShopping при редактировании: одинаковые пропорции, конвертация в один формат, удаление прозрачности, именование файлов. Используется библиотека Intervention Image."
-PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_INSTALL = "Спасибо за установку расширения"
-PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_DISCOVER_INSTALL = "Спасибо за установку расширения"
-PLG_WT_ON_FLY_IMAGE_HANDLER_AFTER_UPDATE = "Спасибо за обновление расширения"
-PLG_WT_ON_FLY_IMAGE_HANDLER_MAYBE_INTERESTING="<h4>Может быть интересно</h4><ul class=\"nav nav-tabs nav-stacked list-group\"><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/action-log-joomshopping.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Action log - JoomShopping - логирование действий пользователя в JoomShopping.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/wt-sm-ozon-rocket-dlya-joomshopping.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT SM Ozon Rocket для JoomShopping, расчет стоимости доставки.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/wt-sm-otpravka-pochta-ru.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Интеграция JoomShopping с Почтой России для Бизнеса</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/wt-joomshopping-favorite.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT JoomShopping Favorites, альтернативный список пожеланий (wishlist) для JoomShopping.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins/wt-joomshopping-schema-org.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT JoomShopping Schema.Org, плагин добавляет на страницы товара и категории товара микроразметку Schema.org для категорий товаров, производителей и товаров JoomShopping в формате ld+json</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomshopping/shablon-joomshopping-na-bootstrap-4-6.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">Шаблон Joomshopping на Bootstrap 4.6.</a></li><li class=\"list-group-item\"><a href=\"https://web-tolk.ru/dev/joomla-plugins/wt-joomshopping-bitrix24-pro.html\" target=\"_blank\" title=\"Перейти на https://web-tolk.ru\">WT JoomShopping Bitrix24 PRO, плагин двухсторонней интеграции интернет-магазина JoomShopping и CRM Битрикс24.</a></li></ul>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_WHATS_NEW="<h3>v.1.0.2 Что нового?</h3><p>Поддержка Joomla 4 и JoomShopping 5.</p><h4>Демо видео</h4><iframe width=\"560\" height=\"315\" src=\"https://www.youtube.com/embed/6_XOuLCy2ng\" title=\"YouTube video player\" frameborder=\"0\" allow=\"accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture\" allowfullscreen></iframe></p>"
-
-PLG_WT_ON_FLY_IMAGE_HANDLER_DOWNLOAD_JINTERVENTION = "<a href=\"https://hika.su/update/free/lib_jinterventionimage.zip\" class=\"btn btn-success\">Скачать библиотеку JIntervention Image [1]</a> <a href=\"https://web-tolk.ru/dev/biblioteki/jintervention-image.html\" class=\"btn btn-success\">Скачать библиотеку JIntervention Image [2]</a>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_ENABLE_JINTERVENTION = "<a href=\"index.php?option=com_installer&view=manage&filter.search=jintervention\" target=\"_blank\" class=\"btn btn-warning\">Включить библиотеку JIntervention Image</a>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_ENABLED="JIntervention Image включена"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_DISABLED="JIntervention Image <span class=\"badge badge-important bg-danger\">выключена</span>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_STATUS_NOT_FOUND="JIntervention Image <span class=\"badge badge-important bg-danger\">не установлена</span>"
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_FAILED="Не удалось скачать и установить библиотеку JIntervention Image автоматически. Пожалуйста, установите её вручную."
-PLG_WT_ON_FLY_IMAGE_HANDLER_JINTERVENTION_DOWNLOAD_OK="Библиотека JIntervention Image установлена автоматически."
