+++ services/provider.php
@@ -1,0 +1,37 @@
+<?php
+
+defined('_JEXEC') || die;
+
+use Joomla\CMS\Extension\PluginInterface;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Plugin\PluginHelper;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+use Joomla\Event\DispatcherInterface;
+use Joomla\Plugin\Jshoppingadmin\Wt_on_fly_image_handler\Extension\Wt_on_fly_image_handler;
+
+return new class () implements ServiceProviderInterface {
+    /**
+     * Registers the service provider with a DI container.
+     *
+     * @param   Container  $container  The DI container.
+     *
+     * @return  void
+     *
+     * @since   4.0.0
+     */
+    public function register(Container $container)
+    {
+        $container->set(
+            PluginInterface::class,
+            function (Container $container) {
+                $subject = $container->get(DispatcherInterface::class);
+                $config  = (array) PluginHelper::getPlugin('jshoppingadmin', 'wt_on_fly_image_handler');
+                $plugin =  new Wt_on_fly_image_handler($subject, $config);
+                $plugin->setApplication(Factory::getApplication());
+                $plugin->setDatabase(Factory::getContainer()->get('DatabaseDriver'));
+                return $plugin;
+            }
+        );
+    }
+};
